Week 1/AI Model Comparison Sheet/requirements.txt
openai
boto3
google-genai
python-dotenv
requests
